removes commented code
