+++ frontend/amundsen_application/static/js/ducks/tableMetadata/api/v0.ts
@@ -183,39 +183,39 @@ export function getPreviewData(queryParams: PreviewQueryParams) {
     });
 }
 
-export function getTableLineage(key: string) {
-  const tableQueryParams = getTableQueryParams({ key });
-  return axios({
-    url: `${API_PATH}/get_table_lineage?${tableQueryParams}`,
-    method: 'GET',
-  })
-    .then((response: AxiosResponse<LineageAPI>) => ({
-      data: response.data,
-      status: response.status,
-    }))
-    .catch((e: AxiosError<LineageAPI>) => {
-      const { response } = e;
-      const status = response ? response.status : null;
-      return Promise.reject({ status });
-    });
-}
-
-export function getColumnLineage(key: string, columnName: string) {
-  const tableQueryParams = getTableQueryParams({
-    key,
-    column_name: columnName,
-  });
-  return axios({
-    url: `${API_PATH}/get_column_lineage?${tableQueryParams}`,
-    method: 'GET',
-  })
-    .then((response: AxiosResponse<LineageAPI>) => ({
-      data: response.data,
-      status: response.status,
-    }))
-    .catch((e: AxiosError<LineageAPI>) => {
-      const { response } = e;
-      const status = response ? response.status : null;
-      return Promise.reject({ status });
-    });
-}
+// export function getTableLineage(key: string) {
+//   const tableQueryParams = getTableQueryParams({ key });
+//   return axios({
+//     url: `${API_PATH}/get_table_lineage?${tableQueryParams}`,
+//     method: 'GET',
+//   })
+//     .then((response: AxiosResponse<LineageAPI>) => ({
+//       data: response.data,
+//       status: response.status,
+//     }))
+//     .catch((e: AxiosError<LineageAPI>) => {
+//       const { response } = e;
+//       const status = response ? response.status : null;
+//       return Promise.reject({ status });
+//     });
+// }
+//
+// export function getColumnLineage(key: string, columnName: string) {
+//   const tableQueryParams = getTableQueryParams({
+//     key,
+//     column_name: columnName,
+//   });
+//   return axios({
+//     url: `${API_PATH}/get_column_lineage?${tableQueryParams}`,
+//     method: 'GET',
+//   })
+//     .then((response: AxiosResponse<LineageAPI>) => ({
+//       data: response.data,
+//       status: response.status,
+//     }))
+//     .catch((e: AxiosError<LineageAPI>) => {
+//       const { response } = e;
+//       const status = response ? response.status : null;
+//       return Promise.reject({ status });
+//     });
+// }
