+++ frontend/amundsen_application/static/js/ducks/tableMetadata/index.spec.ts
@@ -1,7 +1,6 @@
 import { testSaga } from 'redux-saga-test-plan';
 
 import {
-  Lineage,
   PreviewData,
   PreviewQueryParams,
   TableMetadata,
@@ -33,13 +32,6 @@ import reducer, {
   initialTableDataState,
   initialState,
   TableMetadataReducerState,
-  getTableLineage,
-  getTableLineageFailure,
-  initialTableLineageState,
-  getTableLineageSuccess,
-  getColumnLineage,
-  getColumnLineageFailure,
-  getColumnLineageSuccess,
 } from './reducer';
 
 import {
@@ -55,10 +47,6 @@ import {
   updateColumnDescriptionWorker,
   getPreviewDataWatcher,
   getPreviewDataWorker,
-  getTableLineageWatcher,
-  getTableLineageWorker,
-  getColumnLineageWatcher,
-  getColumnLineageWorker,
 } from './sagas';
 
 import {
@@ -68,8 +56,6 @@ import {
   GetColumnDescription,
   UpdateColumnDescription,
   GetPreviewData,
-  GetTableLineage,
-  GetColumnLineage,
 } from './types';
 
 describe('tableMetadata ducks', () => {
@@ -82,7 +68,6 @@ describe('tableMetadata ducks', () => {
   let testKey: string;
   let testIndex: string;
   let testSource: string;
-  let testLineage: Lineage;
 
   let columnName: string;
   let columnIndex: number;
@@ -113,22 +98,6 @@ describe('tableMetadata ducks', () => {
     testKey = 'tableKey';
     testIndex = '3';
     testSource = 'search';
-    testLineage = {
-      upstream_entities: [
-        {
-          badges: [],
-          cluster: 'cluster',
-          database: 'database',
-          key: 'key',
-          level: 1,
-          name: 'name',
-          schema: 'schema',
-          usage: 100,
-        },
-      ],
-      downstream_entities: [],
-    };
-
     columnIndex = 2;
     columnName = 'column_name';
     emptyPreviewData = {
@@ -287,60 +256,6 @@ describe('tableMetadata ducks', () => {
       const { payload } = action;
       expect(action.type).toBe(GetPreviewData.SUCCESS);
       expect(payload.data).toBe(previewData);
-      expect(payload.status).toBe(status);
-    });
-
-    it('getTableLineage - returns the action to get table lineage', () => {
-      const action = getTableLineage(testKey);
-      const { payload } = action;
-      expect(action.type).toBe(GetTableLineage.REQUEST);
-      expect(payload.key).toBe(testKey);
-    });
-
-    it('getTableLineage - returns the action to process failure', () => {
-      const status = 500;
-      const action = getTableLineageFailure(status);
-      const { payload } = action;
-      expect(action.type).toBe(GetTableLineage.FAILURE);
-      expect(payload.lineage).toBe(initialTableLineageState.lineage);
-      expect(payload.status).toBe(status);
-    });
-
-    it('getTableLineage - returns the action to process success', () => {
-      const status = 200;
-      const action = getTableLineageSuccess(testLineage, status);
-      const { payload } = action;
-      expect(action.type).toBe(GetTableLineage.SUCCESS);
-      expect(payload.lineage).toBe(testLineage);
-      expect(payload.status).toBe(status);
-    });
-
-    it('getColumnLineage - returns the action to get column lineage', () => {
-      const action = getColumnLineage(testKey, columnName);
-      const { payload, meta } = action;
-      expect(action.type).toBe(GetColumnLineage.REQUEST);
-      expect(payload.key).toBe(testKey);
-      expect(payload.columnName).toBe(columnName);
-      expect(meta.analytics).not.toBeNull();
-    });
-
-    it('getColumnLineage - returns the action to process failure', () => {
-      const status = 500;
-      const action = getColumnLineageFailure(columnName, status);
-      const { payload } = action;
-      expect(action.type).toBe(GetColumnLineage.FAILURE);
-      expect(payload.columnName).toBe(columnName);
-      expect(payload.lineage).toBe(initialTableLineageState.lineage);
-      expect(payload.status).toBe(status);
-    });
-
-    it('getColumnLineage - returns the action to process success', () => {
-      const status = 200;
-      const action = getColumnLineageSuccess(testLineage, columnName, status);
-      const { payload } = action;
-      expect(action.type).toBe(GetColumnLineage.SUCCESS);
-      expect(payload.columnName).toBe(columnName);
-      expect(payload.lineage).toBe(testLineage);
       expect(payload.status).toBe(status);
     });
   });
@@ -782,72 +697,6 @@ describe('tableMetadata ducks', () => {
           // @ts-ignore TODO: Investigate why redux-saga-test-plan throw() complains
           .throw({ data: previewData, status: 500 })
           .put(getPreviewDataFailure(previewData, 500))
-          .next()
-          .isDone();
-      });
-    });
-
-    describe('getTableLineageWatcher', () => {
-      it('takes every GetTableLineage.REQUEST with getTableLineageWorker', () => {
-        testSaga(getTableLineageWatcher)
-          .next()
-          .takeEvery(GetTableLineage.REQUEST, getTableLineageWorker)
-          .next()
-          .isDone();
-      });
-    });
-
-    describe('getTableLineageWorker', () => {
-      it('executes flow for getting table lineage', () => {
-        testSaga(getTableLineageWorker, getTableLineage(testKey))
-          .next()
-          .call(API.getTableLineage, testKey)
-          .next({ data: testLineage, status: 200 })
-          .put(getTableLineageSuccess(testLineage, 200))
-          .next()
-          .isDone();
-      });
-
-      it('handles request error', () => {
-        testSaga(getTableLineageWorker, getTableLineage(testKey))
-          .next()
-          .call(API.getTableLineage, testKey)
-          // @ts-ignore
-          .throw({ status: 500 })
-          .put(getTableLineageFailure(500))
-          .next()
-          .isDone();
-      });
-    });
-
-    describe('getColumnLineageWatcher', () => {
-      it('takes every GetColumnLineage.REQUEST with getColumnLineageWorker', () => {
-        testSaga(getColumnLineageWatcher)
-          .next()
-          .takeEvery(GetColumnLineage.REQUEST, getColumnLineageWorker)
-          .next()
-          .isDone();
-      });
-    });
-
-    describe('getColumnLineageWorker', () => {
-      it('executes flow for getting column lineage', () => {
-        testSaga(getColumnLineageWorker, getColumnLineage(testKey, columnName))
-          .next()
-          .call(API.getColumnLineage, testKey, columnName)
-          .next({ data: testLineage, status: 200 })
-          .put(getColumnLineageSuccess(testLineage, columnName, 200))
-          .next()
-          .isDone();
-      });
-
-      it('handles request error', () => {
-        testSaga(getColumnLineageWorker, getColumnLineage(testKey, columnName))
-          .next()
-          .call(API.getColumnLineage, testKey, columnName)
-          // @ts-ignore
-          .throw({ status: 500 })
-          .put(getColumnLineageFailure(columnName, 500))
           .next()
           .isDone();
       });
