+++ frontend/amundsen_application/static/js/ducks/tableMetadata/types.ts
@@ -1,5 +1,4 @@
 import {
-  AnalyticsEvent,
   DashboardResource,
   OwnerDict,
   PreviewData,
@@ -7,7 +6,6 @@ import {
   TableMetadata,
   UpdateOwnerPayload,
   Tag,
-  Lineage,
 } from 'interfaces';
 
 export enum GetTableData {
@@ -154,48 +152,5 @@ export interface UpdateTableOwnerResponse {
   type: UpdateTableOwner.SUCCESS | UpdateTableOwner.FAILURE;
   payload: {
     owners: OwnerDict;
-  };
-}
-
-export enum GetTableLineage {
-  REQUEST = 'amundsen/tableMetadata/GET_TABLE_LINEAGE_REQUEST',
-  SUCCESS = 'amundsen/tableMetadata/GET_TABLE_LINEAGE_SUCCESS',
-  FAILURE = 'amundsen/tableMetadata/GET_TABLE_LINEAGE_FAILURE',
-}
-export interface GetTableLineageRequest {
-  type: GetTableLineage.REQUEST;
-  payload: {
-    key: string;
-  };
-}
-export interface GetTableLineageResponse {
-  type: GetTableLineage.SUCCESS | GetTableLineage.FAILURE;
-  payload: {
-    lineage: Lineage;
-    status: number;
-  };
-}
-
-export enum GetColumnLineage {
-  REQUEST = 'amundsen/tableMetadata/GET_COLUMN_LINEAGE_REQUEST',
-  SUCCESS = 'amundsen/tableMetadata/GET_COLUMN_LINEAGE_SUCCESS',
-  FAILURE = 'amundsen/tableMetadata/GET_COLUMN_LINEAGE_FAILURE',
-}
-export interface GetColumnLineageRequest {
-  type: GetColumnLineage.REQUEST;
-  payload: {
-    key: string;
-    columnName: string;
-  };
-  meta: {
-    analytics: AnalyticsEvent;
-  };
-}
-export interface GetColumnLineageResponse {
-  type: GetColumnLineage.SUCCESS | GetColumnLineage.FAILURE;
-  payload: {
-    lineage: Lineage;
-    columnName: string;
-    status: number;
   };
 }
