+++ frontend/amundsen_application/static/js/interfaces/index.ts
@@ -9,3 +9,4 @@ export * from './Resources';
 export * from './TableMetadata';
 export * from './Tags';
 export * from './User';
+export * from './Lineage';
